Showing hovered entity's name in the Editor
- also I havent tested it yet, but probably fixed a bug, where in the HierarchyPanel we were displaying a Texture casting its handle to a uint32_t, thus making it work with opengl, but probably the Vulkna error was because of it

+++ KerberosEditor/src/Editor/HierarchyPanel.cpp
@@ -635,7 +635,7 @@ namespace Kerberos
 				{
 					ImGui::Separator();
 					ImGui::Text("Texture");
-					const uint32_t textureID = staticMesh.MeshTexture->GetRendererID();
+					const uint64_t textureID = staticMesh.MeshTexture->GetRendererID();
 					ImGui::Image(textureID, ImVec2{ 64, 64 }, ImVec2{ 0, 1 }, ImVec2{ 1, 0 });
 				}
